Hominid Mammary Gland Physics Assets
Copy and pasted the physics objects involved in rendering the hominid mammary glands as showcased in 'On the Hypothetical Physics of Hominid Mammary Glands Rendered in a Euclidean Plane Through Application of Verlet Integration'.

This system will be rewritten in the future, but currently serves for ease of use in development.