Primitive Cake Rendering
The entire bakery is currently rendered on the player select screen. No physics is applied yet.

ISSUES: There is a memory leak with the current method of rendering. This needs to be fixed.

The bakery is not yet rendering in world yet, just on the player select screen for some reason. This needs to be investigated.

Physics has not yet been applied yet. Ideally, physics should be applied while in world while 'ideal rendering' is applied on the player select screen. BigBootyParticle will need to be amended to accomodate this.