Better Cake Display -- Leg Draw Interaction
Removed the LegData fetcher's reliance on sets and instead assumes the first layer added during the DrawLeg method are, in fact, the legs (as they usually are).

The cake now correctly depends on the legging color as its source of coloring. This makes it slightly more performant as well as handles the opacity that derives from hit cooldowns.

Added more cake display conditions that gel better with leg drawing.