Removed Slight Jiggle Physics
Considering the price to implement jiggle physics vs the product on screen, this branch has opted to remove the physics while retaining only the 'sprite-sheet' style rendering of the dump truck parking lot.

Consequently, BigBootyParticles is now a significantly smaller struct (as opposed to a class in the physics branch which, granted, should likely be redesigned into a struct).